pyproject.toml
[tool.poetry]
name = "py_linear"
version = "0.1.0"
description = "A Linear CLI"
authors = ["TumuGuskun <johntgaskin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
gql-query-builder = "^0.1.7"
aiohttp = "^3.8.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
greet = "greet.location:cli"
